+++ thrift/src/main/scala/com/twitter/finatra/thrift/ThriftServer.scala
@@ -2,6 +2,7 @@ package com.twitter.finatra.thrift
 
 import com.google.inject.Module
 import com.twitter.conversions.time._
+import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.{ListeningServer, NullServer, ThriftMux}
 import com.twitter.finatra.thrift.modules.{ExceptionManagerModule, ThriftResponseClassifierModule}
 import com.twitter.finatra.thrift.response.ThriftResponseClassifier
@@ -75,6 +76,7 @@ trait ThriftServer extends TwitterServer {
       configureThriftServer(
         ThriftMux.server
           .withLabel(thriftServerNameFlag())
+          .withStatsReceiver(injector.instance[StatsReceiver].scope("srv"))
           .withResponseClassifier(injector.instance[ThriftResponseClassifier])
       )